Add custom TermTree
