+++ examples/text_to_knowledge/termtree/README.md
@@ -17,6 +17,37 @@ TermTree（百科知识树）是一个描述所有中文词汇（包括概念、
 
 **注：** 与其他常见应用知识图谱不同，TermTree的核心是概念词，而非专名实体词。因为，在中文文本中，概念词的含义是相对稳定的，而专名实体词随应用变化（例如，不同电商有不同的商品实体集，不同的小说站有不同的小说实体集），因此，TermTree通过 “提供常用概念集 + 可插拔的应用实体集/应用知识图谱” 来达到支持不同的应用适配。
 
+## 自定义TermTree
+
+`termtree.py`文件中的TermTree类支持TermTree的加载、增加、保存操作，因为涉及到数据结构整体性和一致性，暂不支持删除和修改操作。下面提供了离线维护自定义TermTree的代码示例
+
+### 文件准备
+
+首先下载已有的TermTreeV1.0
+```shell
+wget https://kg-concept.bj.bcebos.com/TermTree/TermTree.V1.0.tar.gz && tar -zxvf TermTree.V1.0.tar.gz
+```
+
+### TermTree维护与修改
+
+加载TermTreeV1.0，增加新的term
+```shell
+from termtree import TermTree
+
+# 加载百科知识树
+termtree = TermTree.from_dir("termtree_type.csv", "TermTree.V1.0")
+
+# 增加新term: 平原上的火焰
+termtree.add_term(term="平原上的火焰",
+                  base="eb",
+                  term_type="影视作品")
+
+# 保存修改, 执行后将在当前路径生成文件`termtree_data`，即新的自定义TermTree
+termtree.save("./")
+```
+
+## 常见问题
+
 **常见问题1：为什么TermTree采用树状结构（Tree），而不是网状结构（Net/Graph）？**
 
 - 树结构是对知识空间的全划分，网状结构是对相关关系的描述和提炼。树结构更方便做到对词类体系的全面描述。
