+++ examples/text_to_knowledge/termtree/termtree_type.csv
@@ -1,163 +1,164 @@
-﻿type-1,type-2,type-3,说明,主要对应词性,subtype示例(可扩展的开放集合),src(C表示cb基础概念库、E表示eb基础实体库)
-角色,,,角色类概念和实体,n/普通名词;nr/人名,群体,C/E
-角色,人物,,人物类概念和实体,nr/人名,职业角色、历史人物、行业人物,C/E
-角色,民族族群,,民族和族群,,五十六个民族,C/E
-角色,虚拟角色,,非现实的角色,,虚拟人物、虚拟生物,C/E
-作品与出版物,,,作品类概念和实体,nw/作品名,拓片,C/E
-作品与出版物,游戏,,游戏类概念和实体,,电子游戏、视频小游戏、网页游戏,C/E
-作品与出版物,影视动漫作品,,包括各类视频,,视频作品,C/E
-作品与出版物,影视动漫作品,动漫作品,动漫作品类概念和实体,,漫画、动画,C/E
-作品与出版物,影视动漫作品,影视作品,影视作品类概念和实体,,电影、电视剧,C/E
-作品与出版物,影视动漫作品,视频节目,视频节目类概念和实体,,脱口秀节目、新闻类节目、访谈节目,C/E
-作品与出版物,音乐,,音乐类概念和实体,,歌曲、音乐专辑,C/E
-作品与出版物,小说,,小说类概念和实体,,网络小说、言情小说,C/E
-作品与出版物,诗歌,,诗歌作品类概念和实体,,诗词,C/E
-作品与出版物,计算机软件,,计算机软件类概念和实体,,工具软件、办公软件,C/E
-作品与出版物,舞蹈,,舞蹈类概念和实体,,,C/E
-作品与出版物,美术,,美术类概念和实体,,雕塑作品、油画作品、工艺美术作品,C/E
-作品与出版物,图书,,图书类概念和实体,,书籍、词典、教材,C/E
-作品与出版物,刊物,,刊物类概念和实体,,报纸、期刊,C/E
-作品与出版物,文件,,各类文件和文书,,文书,C/E
-区域场所,,,区域场所类,ns/地名,宗教场所、建筑物、地域范围,C/E
-区域场所,景点,,景点类概念和实体,,公园、植物园、动物园、博物馆,C/E
-区域场所,楼盘住宅,,楼盘住宅类概念和实体,,商业楼盘、住宅楼盘、住宅小区,C/E
-区域场所,交通场所,,交通场所类概念和实体,,机场、车站、港口、交通道路、交通线路,C/E
-区域场所,住宿场所,,住宿场所类概念和实体,,酒店、旅馆,C/E
-区域场所,餐饮场所,,餐饮场所类概念和实体,,咖啡馆、餐馆,C/E
-区域场所,网上组织机构场所,,网站等概念和实体,,网站、虚拟社区,C/E
-位置方位,,,位置方位词;便于区分区域场所,f/方位名词;s/处所名词,,C
-组织机构,,,组织机构类概念和实体,nt/机构团体名,委员会、论坛、组织机构范围,C/E
-组织机构,演艺团体,,演艺团体类概念和实体,,乐队、艺术团、偶像组合,C/E
-组织机构,国家机关,,国家机关类概念和实体,,政府部门、党政机关,C/E
-组织机构,企事业单位,,企业公司类概念和实体,,公司、厂商、企业,C/E
-组织机构,教育组织机构,,教育组织机构类概念和实体,,学校、大学、幼儿园、培训机构,C/E
-组织机构,居民服务机构,,居民服务机构类概念和实体,,母婴护理机构、婚介机构、美容护理机构、家政服务机构,C/E
-组织机构,医疗卫生机构,,医疗卫生机构类概念和实体,,医院、药店、诊所、科室,C/E
-组织机构,体育组织机构,,体育组织机构类概念和实体,,运动队、体育俱乐部,C/E
-组织机构,金融组织机构,,金融组织机构类概念和实体,,银行、交易所、投资机构,C/E
-组织机构,军事组织机构,,军事组织机构类概念和实体,,部队、军区,C/E
-品牌,,,品牌类概念和品牌名,n/普通名词;nz/其他专名,,C/E
-品牌,汽车品牌,,"汽车品牌;从""汽车""类筛选;便于领域应用",,,C/E
-品牌,手机品牌,,"手机品牌;从""手机""类筛选;便于领域应用",,,C/E
-品牌,个护用品品牌,,个护用品品牌,,护肤品牌、彩妆品牌,C/E
-物体与物品,,,包括物品和物质,n/普通名词;nz/其他专名,物体构造、化学物质,C/E
-物体与物品,物品,,物品类概念和实体,,飞机、船舶、轴承、摄影器材,C/E
-物体与物品,物品,汽车,汽车类概念和实体,,,C/E
-物体与物品,物品,手机,手机类概念和实体,,,C/E
-物体与物品,物品,美容美发用品,美容美发用品类概念和实体,,化妆品、美发用品,C/E
-物体与物品,物品,电子电器产品,电子电器产品类概念和实体,,计算机、家用电器,C/E
-物体与物品,物品,衣物饰品,衣物饰品类概念和实体,,服装、箱包、鞋靴、饰品配件,C/E
-物体与物品,物品,兵器,兵器类概念和实体,,武器、导弹、冷兵器..,C/E
-物体与物品,设施,,设施类概念和实体,,,C/E
-饮食,,,饮食类概念和实体,n/普通名词;nz/其他专名,食材,C/E
-饮食,菜品,,菜品类概念和实体,,汤品、面食,C/E
-饮食,饮品,,饮品类概念和实体,,茶叶、酒、饮料和冷饮类,C/E
-生物,,,生物物种,n/普通名词;nz/其他专名,,C
-生物,动物,,动物物种,,猫、狗、鸟纲、昆虫纲、鱼纲,C
-生物,植物,,植物物种,,,C
-生物,微生物,,微生物物种,,真菌、细菌、生物病毒,C
-世界地区,,,世界地区;包括地球外区域,ns/地名,首都、地球山脉、地球河流、地球岛屿、历史地区,C
-世界地区,中国地区,,中国地区,,中国省级行政区、中国省会,C
-世界地区,国家,,现代意义的国家,,现存国家、历史国家,C
-虚拟事物,,,非现实事物,n/普通名词;nz/其他专名,虚拟场所、虚拟场景,C/E
-虚拟事物,虚拟物品,,虚拟物品;如游戏物品等,,虚拟宝物、游戏卡牌、游戏道具、游戏装备,C/E
-文化,,,文化相关的特定类目,n/普通名词;nz/其他专名,,C/E
-文化,姓氏与人名,,姓氏和名字（非特定人物）,,中文姓氏、英文姓氏,C
-文化,语言文字,,语言、语系、方言等,,汉语、方言,C
-文化,民俗,,算命占星等常用词,,方术、数术、十二生肖、占星学星座、周易六十四卦,C/E
-文化,政权朝代,,包括古代政权、古代国家,,历史政权、中国朝代,C/E
-文化,制度政策协议,,制度、政策、协议等,,,C/E
-文化,奖项赛事活动,,奖项、赛事、活动等,,奖项、体育赛事,C/E
-事件,,,事件类,n/普通名词;nz/其他专名,展览、会议、案件、事故、战争,C/E
-术语,,,领域术语类;收录生活或百科中常见类目;可根据应用需要扩展,n/普通名词;nz/其他专名,,C/E
-术语,编码符号指标,,,,价格、符号、信号、度量单位、邮政编码.....,C
-术语,教育用语,,,,,C
-术语,教育用语,学科,,,,C
-术语,教育用语,学历学位,,,学历、学位,C
-术语,教育用语,专业,,,,C
-术语,游戏用语,,,,,C
-术语,游戏用语,麻将术语,,,,C
-术语,医药学术语,,,,中医术语、西医术语、医学指标、诊断治疗方法,C
-术语,医药学术语,医疗美容项目,,,,C
-术语,医药学术语,药物,,,中药、西药,C/E
-术语,医药学术语,疾病损伤,,,疾病、疾病症状,C
-术语,医药学术语,动物疾病,,,,C
-术语,金融术语,,,,股票术语、证券术语、保险术语、银行术语,C
-术语,金融术语,股票,,,,C/E
-术语,金融术语,保险,,,,C/E
-术语,金融术语,基金,,,,C/E
-术语,金融术语,银行卡,,,借记卡、信用卡,C/E
-术语,经济术语,,,,会计术语,C
-术语,法律术语,,,,,C
-术语,法律术语,法律法规,,,法律体系、法律、法规,C/E
-术语,法律术语,罪名,,,,C
-术语,体育术语,,,,围棋术语、象棋术语、篮球术语,C
-术语,体育术语,体育运动项目,,,球类运动、武术功夫,C
-术语,赌博博彩用语,,,,赌博用语,C
-术语,赌博博彩用语,彩票,,,,C
-术语,天文学术语,,,,星系、恒星,C/E
-术语,天文学术语,星座,,,八十八星座,C
-术语,天文学术语,星体,,,小行星,C/E
-术语,生物学术语,,,,,C
-术语,生物学术语,动物体构造,,,动物器官系统、骨,C
-术语,生物学术语,植物病虫害,,,植物病害、植物虫害,C
-术语,机械工程术语,,,,机械制造术语,C
-术语,机械工程术语,汽车术语,,,,C
-术语,大气科学术语,,,,气象学术语、气候学术语,C
-术语,大气科学术语,台风,,,,C/E
-术语,计算机术语,,,,病毒程序、计算机网络术语、编程技术术语,C
-术语,文化术语,,,,摄影术语、音乐术语、文学术语,C
-术语,数学术语,,,,数学概念、数学公式、几何学术语,C
-术语,物理术语,,,,电学术语、力学术语,C
-术语,化学术语,,,,化学结构,C
-术语,统计术语,,,,数理统计术语,C
-术语,地学术语,,,,地理学术语、地质学术语,C
-术语,农业学术语,,,,土壤学术语,C
-术语,心理学术语,,,,心理现象,C
-术语,语言学术语,,,,语法、词法、音韵学术语,C
-术语,建筑术语,,,,土木工程术语、装修术语,C
-术语,军事术语,,,,,C
-术语,政治术语,,,,,C
-术语,哲学术语,,,,哲学理论、伦理学术语、逻辑学术语,C
-术语,宗教术语,,,,道教术语、佛教术语,C
-术语,通信术语,,,,,C
-术语,材料科学术语,,,,,C
-术语,航空科技术语,,,,,C
-术语,水利科技术语,,,,水利工程,C
-术语,测绘学术语,,,,测量术语,C
-术语,电力术语,,,,,C
-术语,社会学术语,,,,,C
-术语,交通术语,,,,船舶工程术语,C
-术语,钓鱼术语,,,,,C
-术语,ACGN术语,,,,,C
-生活用语,,,收录日常生活中常用词,n/普通名词;nz/其他专名,信息知识资料、标识物、行业、服务,C/E
-生活用语,情绪,,收录常见情绪类词,,,C
-生活用语,态度,,收录常见态度类词,,,C
-生活用语,表情,,收录常见表情类词,,笑、哭、眼神,C
-生活用语,人物造型,,收录常见人物造型类词,,妆容、发型,C
-生活用语,个性特点,,收录常见个性特点类词,,,C
-生活用语,颜色,,收录常见颜色类词,,,C
-生活用语,场景事件,,收录常见动词,v/普通动词;vn/名动词;vd/动副词,考试,C/E
-时间阶段,,,时间相关词,t/时间名词,时间、年代、世纪,C
-时间阶段,地质年代,,地质年代,,,C
-时间阶段,特殊日,,节假日、活动纪念日等,,农历二十四节气、假日、节日、纪念日,C
-词汇用语,,,语法词类、汉字、成语等;用于兜底识别,n/普通名词,,C
-词汇用语,汉字,,汉字字表,,,C/E
-词汇用语,成语,,成语词表,,,C/E
-词汇用语,俗语,,非成语的俗语;可根据需求扩充词表,,歇后语、顺口溜、谚语,C/E
-词汇用语,诗句,,诗句;可根据需求扩充词表,,,C/E
-词汇用语,介词,,介词;封闭集合,p/介词,,C
-词汇用语,助词,,助词;封闭集合,u/助词,,C
-词汇用语,代词,,代词;封闭集合,r/代词,,C
-词汇用语,连词,,连词;封闭集合,c/连词,,C
-词汇用语,副词,,副词;封闭集合,d/副词,,C
-词汇用语,疑问词,,疑问词;封闭集合,,,C
-词汇用语,肯定否定词,,常用肯定词和否定词;便于区分场景事件;封闭集合,v/普通动词,,C
-词汇用语,量词,,量词;封闭集合,q/量词,,C
-词汇用语,数量词,,数量词;收录常见词,m/数量词,,C
-词汇用语,叹词,,叹词;封闭集合,,,C
-词汇用语,拟声词,,拟声词;封闭集合,,,C
-词汇用语,修饰词,,修饰词;收录常见形容词;可根据需求扩充词表,n/普通名词;a/形容词;ad/副形词;an/名形词,,C
-词汇用语,汉字偏旁部首,,汉字偏旁部首,,,C
-词汇用语,日文假名,,日文假名,,平假名、片假名,C
-词汇用语,汉语拼音,,汉语拼音字母,,,C
+type-1	type-2	type-3
+世界地区		
+世界地区	中国地区	
+世界地区	国家	
+区域场所		
+区域场所	住宿场所	
+区域场所	楼盘住宅	
+区域场所	餐饮场所	
+区域场所	交通场所	
+区域场所	景点	
+区域场所	网上组织机构场所	
+组织机构		
+组织机构	居民服务机构	
+组织机构	金融组织机构	
+组织机构	体育组织机构	
+组织机构	教育组织机构	
+组织机构	国家机关	
+组织机构	企事业单位	
+组织机构	演艺团体	
+组织机构	医疗卫生机构	
+组织机构	军事组织机构	
+饮食		
+饮食	饮品	
+饮食	菜品	
+品牌		
+品牌	个护用品品牌	
+品牌	汽车品牌	
+品牌	手机品牌	
+文化		
+文化	奖项赛事活动	
+文化	姓氏与人名	
+文化	政权朝代	
+文化	民俗	
+文化	制度政策协议	
+文化	语言文字	
+虚拟事物		
+虚拟事物	虚拟物品	
+位置方位		
+作品与出版物		
+作品与出版物	计算机软件	
+作品与出版物	小说	
+作品与出版物	作品IP	
+作品与出版物	诗歌	
+作品与出版物	图书	
+作品与出版物	音乐	
+作品与出版物	舞蹈	
+作品与出版物	文件	
+作品与出版物	美术	
+作品与出版物	刊物	
+作品与出版物	影视动漫作品	
+作品与出版物	影视动漫作品	动漫作品
+作品与出版物	影视动漫作品	视频节目
+作品与出版物	影视动漫作品	影视作品
+作品与出版物	游戏	
+时间阶段		
+时间阶段	特殊日	
+时间阶段	地质年代	
+生物		
+生物	微生物	
+生物	植物	
+生物	动物	
+物体与物品		
+物体与物品	设施	
+物体与物品	物品	
+物体与物品	物品	电子电器产品
+物体与物品	物品	兵器
+物体与物品	物品	衣物饰品
+物体与物品	物品	手机
+物体与物品	物品	美容美发用品
+物体与物品	物品	汽车
+事件		
+术语		
+术语	医药学术语	
+术语	医药学术语	疾病损伤
+术语	医药学术语	动物疾病
+术语	医药学术语	药物
+术语	医药学术语	医疗美容项目
+术语	化学术语	
+术语	军事术语	
+术语	建筑术语	
+术语	经济术语	
+术语	政治术语	
+术语	交通术语	
+术语	教育用语	
+术语	教育用语	学科
+术语	教育用语	专业
+术语	教育用语	学历学位
+术语	体育术语	
+术语	体育术语	体育运动项目
+术语	农业学术语	
+术语	法律术语	
+术语	法律术语	罪名
+术语	法律术语	法律法规
+术语	测绘学术语	
+术语	生物学术语	
+术语	生物学术语	动物体构造
+术语	生物学术语	植物病虫害
+术语	水利科技术语	
+术语	赌博博彩用语	
+术语	赌博博彩用语	彩票
+术语	宗教术语	
+术语	电力术语	
+术语	钓鱼术语	
+术语	社会学术语	
+术语	天文学术语	
+术语	天文学术语	星体
+术语	天文学术语	星座
+术语	语言学术语	
+术语	机械工程术语	
+术语	机械工程术语	汽车术语
+术语	航空科技术语	
+术语	文化术语	
+术语	统计术语	
+术语	数学术语	
+术语	金融术语	
+术语	金融术语	基金
+术语	金融术语	股票
+术语	金融术语	保险
+术语	金融术语	银行卡
+术语	大气科学术语	
+术语	大气科学术语	台风
+术语	计算机术语	
+术语	ACGN术语	
+术语	编码符号指标	
+术语	地学术语	
+术语	心理学术语	
+术语	哲学术语	
+术语	材料科学术语	
+术语	游戏用语	
+术语	游戏用语	麻将术语
+术语	物理术语	
+术语	通信术语	
+生活用语		
+生活用语	情绪	
+生活用语	表情	
+生活用语	个性特点	
+生活用语	场景事件	
+生活用语	态度	
+生活用语	颜色	
+生活用语	人物造型	
+角色		
+角色	人物	
+角色	虚拟角色	
+角色	民族族群	
+词汇用语		
+词汇用语	代词	
+词汇用语	诗句	
+词汇用语	汉字偏旁部首	
+词汇用语	助词	
+词汇用语	修饰词	
+词汇用语	汉语拼音	
+词汇用语	日文假名	
+词汇用语	副词	
+词汇用语	拟声词	
+词汇用语	俗语	
+词汇用语	量词	
+词汇用语	叹词	
+词汇用语	汉字	
+词汇用语	肯定否定词	
+词汇用语	介词	
+词汇用语	数量词	
+词汇用语	疑问词	
+词汇用语	连词	
+词汇用语	成语
